Tag prediction class
I added a new file tags.txt containing all the names and dates of the festivals. And I created a new class to load all the festivals and predict. After that I made some changes to PredictModel::predictCategoriesAndTags to make it be able to predict tags.
Finally I found that date.getMonth() returns a number in [0, 11] which I thought would be [1, 12] and caused some bugs.
About the problem of the interval is not yet fixed.