f**k thread
The MONT_COUNT is set to 1024 and the benchmark class is also changed.
4 classes are refactored from /coursework/ to /coursework/algo/. And solved some of the problem caused by refactoring.
All the predict function in class PredictModel are set to be multithreading. And f**k multithreading, I took 6 hours to realize that there is no lock.
I find that when input transaction list is small, GMM model may still behave poor and some of the gauss distribution may have negative weights, this cause dead loop when calculating models. So finally I decided that I will force all the output to be positive. This may cause that when calculating GMM models, the sum of all the weights is bigger than 1, so when calculating the models, all the weights must divided by their sum.
I used commonMath.KMeansPlusPlusClusterer to replace my hand-written KMeans.